refactor(strategy): 기물 배치 전략 구조 개선 및 Resolver 도입
- 최상위 ArrangementStrategy를 인터페이스로 분리해 역할 명확화
- 마/상 배치를 제외한 고정 기물 배치를 담당하는 MaSangArrangementTemplate 추상 클래스 도입: 기물
초기화시 발생하던 덮어쓰기 문제 해결 및 OCP 확보
- 도메인이 UI 데이터를 들고 있던 StrategyLabel을 삭제하고, 입력 번호와 전략 객체를 매핑하는
ArrangementStrategyResolver 도입
- JanggiFlow 내부의 전략 객체 매핑 및 캐싱 로직을 Resolver로 위임해 컨트롤러 무상태 유지 및 책임 분리
- 하위 전략 구현체들의 네이밍을 명확하게 변경 (ex. MaSangMaSang -> MaSangMaSangStrategy)
