+++ src/main/java/janggi/strategy/SangMaMaSangStrategy.java
@@ -6,15 +6,11 @@ import janggi.domain.piece.Piece;
 import janggi.domain.piece.Sang;
 
 @SuppressWarnings("java:S6548")
-public class MaSangSangMa extends ArrangementStrategy {
+public class SangMaMaSangStrategy extends MaSangArrangementTemplate {
 
-    private static final MaSangSangMa INSTANCE = new MaSangSangMa();
+    private static final SangMaMaSangStrategy INSTANCE = new SangMaMaSangStrategy();
 
-    private MaSangSangMa() {
-        super(StrategyLabel.MSSM);
-    }
-
-    public static MaSangSangMa getInstance() {
+    public static SangMaMaSangStrategy getInstance() {
         return INSTANCE;
     }
 
@@ -22,9 +18,9 @@ public class MaSangSangMa extends ArrangementStrategy {
     protected void placeVariablePieces(Piece[][] board, Side side) {
         int boardMaxLength = board.length;
         int row = calculateInitialRow(boardMaxLength, side);
-        board[row][1] = new Ma(side);
-        board[row][2] = new Sang(side);
-        board[row][6] = new Sang(side);
-        board[row][7] = new Ma(side);
+        board[row][1] = new Sang(side);
+        board[row][2] = new Ma(side);
+        board[row][6] = new Ma(side);
+        board[row][7] = new Sang(side);
     }
 }
