+++ src/main/java/janggi/strategy/SangMaSangMaStrategy.java
@@ -6,15 +6,11 @@ import janggi.domain.piece.Piece;
 import janggi.domain.piece.Sang;
 
 @SuppressWarnings("java:S6548")
-public class MaSangMaSang extends ArrangementStrategy {
+public class SangMaSangMaStrategy extends MaSangArrangementTemplate {
 
-    private static final MaSangMaSang INSTANCE = new MaSangMaSang();
+    private static final SangMaSangMaStrategy INSTANCE = new SangMaSangMaStrategy();
 
-    private MaSangMaSang() {
-        super(StrategyLabel.MSMS);
-    }
-
-    public static MaSangMaSang getInstance() {
+    public static SangMaSangMaStrategy getInstance() {
         return INSTANCE;
     }
 
@@ -22,9 +18,9 @@ public class MaSangMaSang extends ArrangementStrategy {
     protected void placeVariablePieces(Piece[][] board, Side side) {
         int boardMaxLength = board.length;
         int row = calculateInitialRow(boardMaxLength, side);
-        board[row][1] = new Ma(side);
-        board[row][2] = new Sang(side);
-        board[row][6] = new Ma(side);
-        board[row][7] = new Sang(side);
+        board[row][1] = new Sang(side);
+        board[row][2] = new Ma(side);
+        board[row][6] = new Sang(side);
+        board[row][7] = new Ma(side);
     }
 }
