+++ src/test/java/janggi/strategy/MaSangSangMaStrategyTest.java
@@ -9,13 +9,13 @@ import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-class MaSangMaSangTest {
+class MaSangSangMaStrategyTest {
 
-    private static final ArrangementStrategy STRATEGY = MaSangMaSang.getInstance();
+    private static final MaSangSangMaStrategy STRATEGY = MaSangSangMaStrategy.getInstance();
 
     @Test
-    @DisplayName("Han 팀의 Sang 객체와 Ma 객체를 MaSangMaSang의 위치에 생성해 넣어준다.")
-    void shouldPlaceMaSangMaSangWhenSideHan() {
+    @DisplayName("Han 팀의 Sang 객체와 Ma 객체를 MaSangSangMa의 위치에 생성해 넣어준다.")
+    void shouldPlaceMaSangSangMaWhenSideHan() {
         // given
         Piece[][] grid = new Piece[10][9];
         Side side = Side.HAN;
@@ -24,24 +24,24 @@ class MaSangMaSangTest {
         STRATEGY.place(grid, side);
         Piece leftMa = grid[0][1];
         Piece leftSang = grid[0][2];
-        Piece rightMa = grid[0][6];
-        Piece rightSang = grid[0][7];
+        Piece rightSang = grid[0][6];
+        Piece rightMa = grid[0][7];
 
         // then
         Assertions.assertThat(leftMa).isInstanceOf(Ma.class);
         Assertions.assertThat(leftSang).isInstanceOf(Sang.class);
-        Assertions.assertThat(rightMa).isInstanceOf(Ma.class);
         Assertions.assertThat(rightSang).isInstanceOf(Sang.class);
+        Assertions.assertThat(rightMa).isInstanceOf(Ma.class);
 
         Assertions.assertThat(leftMa.isSameSide(side)).isTrue();
         Assertions.assertThat(leftSang.isSameSide(side)).isTrue();
-        Assertions.assertThat(rightMa.isSameSide(side)).isTrue();
         Assertions.assertThat(rightSang.isSameSide(side)).isTrue();
+        Assertions.assertThat(rightMa.isSameSide(side)).isTrue();
     }
 
     @Test
-    @DisplayName("Cho 팀의 Sang 객체와 Ma 객체를 MaSangMaSang의 위치에 생성해 넣어준다.")
-    void shouldPlaceMaSangMaSangWhenSideCho() {
+    @DisplayName("Cho 팀의 Sang 객체와 Ma 객체를 MaSangSangMa의 위치에 생성해 넣어준다.")
+    void shouldPlaceMaSangSangMaWhenSideCho() {
         // given
         Piece[][] grid = new Piece[10][9];
         Side side = Side.CHO;
@@ -50,18 +50,18 @@ class MaSangMaSangTest {
         STRATEGY.place(grid, side);
         Piece leftMa = grid[9][1];
         Piece leftSang = grid[9][2];
-        Piece rightMa = grid[9][6];
-        Piece rightSang = grid[9][7];
+        Piece rightSang = grid[9][6];
+        Piece rightMa = grid[9][7];
 
         // then
         Assertions.assertThat(leftMa).isInstanceOf(Ma.class);
         Assertions.assertThat(leftSang).isInstanceOf(Sang.class);
-        Assertions.assertThat(rightMa).isInstanceOf(Ma.class);
         Assertions.assertThat(rightSang).isInstanceOf(Sang.class);
+        Assertions.assertThat(rightMa).isInstanceOf(Ma.class);
 
         Assertions.assertThat(leftMa.isSameSide(side)).isTrue();
         Assertions.assertThat(leftSang.isSameSide(side)).isTrue();
-        Assertions.assertThat(rightMa.isSameSide(side)).isTrue();
         Assertions.assertThat(rightSang.isSameSide(side)).isTrue();
+        Assertions.assertThat(rightMa.isSameSide(side)).isTrue();
     }
 }
